Changed atoi for atoll in GetDiskSpace

+++ file-system-v3.0/lib/file.hpp
@@ -19,7 +19,7 @@ size_t GetDiskSpace() {
   char buffer[5];
   char *line_p = fgets(buffer, sizeof(buffer), command);
   pclose(command);
-  size_t space = atoi(line_p);
+  size_t space = atoll(line_p);
   return space;
 }
